requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.24
Pillow>=10.0
faiss-cpu>=1.7.4
sentence-transformers>=3.0.0
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.30.0
huggingface_hub>=0.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0